fix(scheduling): 处理 PR #180 review 全部反馈（2 Critical + 6 实质 + 2 Nit）
Critical
- #1 RowJobLock 表名不硬编码：从 EF 元数据 GetTableName() 读回，
  显式 ToTable 固定权威名。SQL 路径用 ExecuteSqlRawAsync + 显式
  DbParameter（表名内联受控常量、值参数化），杜绝复数/单数漂移致崩溃。
- #2 白名单 fail-secure：null+JobAssemblies空 → 不限制；显式 [] 或
  不含本程序集 → 拒绝（防反射注入）。

实质
- #3 JobContext.Parameters 改 JsonElement?，加 GetParameters<T>()；
  JobExecutor 解析 ParametersJson 传入，非法 JSON 按 null 不阻断执行。
- #4 BuildJobTypeName 用 Type.AssemblyQualifiedName（含版本/公钥，稳健）。
- #5 JobExecutor 写入/更新路径 IgnoreQueryFilters（后台无请求上下文，
  ITenantContext 解析抛异常）。
- #6 超时 CTS 改局部 using（原实例字段并发执行会互相覆盖/Dispose）。
- #7 默认 JobTimeout=5min / LockLeaseDuration=5min（满足 ≥ 约束），
  AddTenE0Scheduling 加 ValidateOnStart 校验 lease ≥ timeout。
- #8 JobFailedEvent 移到所有 DB 提交之后触发（InProcess dispatcher 非
  Outbox 持久化，提交前发事件会在回滚时丢失一致性）。

Nit
- AddOptions+Configure 合并为单 AddOptions().Configure().Validate()。
- CreateScheduledJobDto.RetryIntervalMs double → long（毫秒整数语义）。

测试 +8（白名单 fail-secure 三态 / 参数反序列化 / lease-timeout 校验），
模块 CLAUDE.md 同步更新。全部 1505 测试通过，0 警告，format 干净。

+++ src/10E0.Api/Endpoints/AdminEndpoints.cs
@@ -616,7 +616,9 @@ internal sealed class CreateScheduledJobDto
     public required string JobType { get; set; }
     public string? ParametersJson { get; set; }
     public int MaxRetries { get; set; } = 3;
-    public double RetryIntervalMs { get; set; } = 60_000;
+
+    // long 而非 double：毫秒是整数单位，long 语义更清晰且避免浮点精度问题。
+    public long RetryIntervalMs { get; set; } = 60_000;
 }
 
 internal sealed class UpdateScheduledJobDto
@@ -626,5 +628,5 @@ internal sealed class UpdateScheduledJobDto
     public bool? IsEnabled { get; set; }
     public string? ParametersJson { get; set; }
     public int? MaxRetries { get; set; }
-    public double? RetryIntervalMs { get; set; }
+    public long? RetryIntervalMs { get; set; }
 }
